Changes needed to use pkg in 2D game
We are dogfooding library by using it in a 2D game engine gousion2D.
Still lacking tests. Lots of functionality still missing. We are building
as little as possible to get stuff up on the screen while at the same time
heading in the right direction with respect to creating a data oriented design
for a game engine.